+++ MovesenseCoreLib/MovesenseFromStaticLib.cmake
@@ -247,7 +247,10 @@ set(SOURCES ${SOURCES} ${MOVESENSE_CORE_LIBRARY}/src/movesense_heap.c)
 
 # Application sources in current folder (app)
 aux_source_directory(. APP_SOURCES)
-set(APP_SOURCES ${APP_SOURCES} ${SOURCES})
+set(APP_SOURCES ${APP_SOURCES} ${SOURCES}
+    ../ble-ecg/offline_storage_client/OfflineStorageGattClient.cpp
+    ../ble-ecg/offline_storage_client/MeasurementProvider.cpp
+)
 
 
 ####################################