+++ ble-ecg/App.cpp
@@ -1,7 +1,6 @@
-// #include "EcgGATTSvcClient.h"
-// #include "MovGATTSvcClient.h"
-#include "ActivityGATTSvcClient.h"
 #include "WakeClient.h"
+#include "offline_storage_client/OfflineStorageGattClient.h"
+#include "offline_storage_client/MeasurementProvider.h"
 
 #include "movesense.h"
 
@@ -9,20 +8,20 @@ MOVESENSE_APPLICATION_STACKSIZE(1024)
 
 // Add Providers:
 
-MOVESENSE_PROVIDERS_BEGIN(2)
+MOVESENSE_PROVIDERS_BEGIN(3)
 
-// MOVESENSE_PROVIDER_DEF(EcgGATTSvcClient)
-MOVESENSE_PROVIDER_DEF(ActivityGATTSvcClient)
+MOVESENSE_PROVIDER_DEF(OfflineStorageGattClient)
+MOVESENSE_PROVIDER_DEF(MeasurementProvider)
 MOVESENSE_PROVIDER_DEF(WakeClient)
 
-MOVESENSE_PROVIDERS_END(2)
+MOVESENSE_PROVIDERS_END(3)
 
 // Load Features:
 
 MOVESENSE_FEATURES_BEGIN()
 
-OPTIONAL_CORE_MODULE(DataLogger, false)
-OPTIONAL_CORE_MODULE(Logbook, false)
+OPTIONAL_CORE_MODULE(DataLogger, true)
+OPTIONAL_CORE_MODULE(Logbook, true)
 OPTIONAL_CORE_MODULE(LedService, true)
 OPTIONAL_CORE_MODULE(IndicationService, true)
 OPTIONAL_CORE_MODULE(BleService, true)
@@ -36,12 +35,10 @@ OPTIONAL_CORE_MODULE(CustomGattService, true)
 
 // Define Application Specification:
 
-APPINFO_NAME("ActivityGATTSvc");
-APPINFO_VERSION("0.5.1");
+APPINFO_NAME("OfflineActivityService");
+APPINFO_VERSION("0.8.0");
 APPINFO_COMPANY("Jonas Primbs");
 
 // Enable Bluetooth Low Energy Communication:
-
-BLE_COMMUNICATION(true)
 
 MOVESENSE_FEATURES_END()